farm/requirements.txt
attrs==23.2.0
autopep8==2.1.0
aws_lambda_powertools==2.37.0
azure-core==1.30.1
azure-eventhub==5.11.7
azure-iot-device==2.13.0
blinker==1.7.0
boto3==1.34.88
botocore==1.34.88
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
cryptography==42.0.5
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
deprecation==2.1.0
evdev==1.7.0
Flask==3.0.3
future==1.0.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gpio==1.0.0
gpiozero==2.0.1
grove==1.2.2
grove.py==0.6
grovepi==1.0.4
httplib2==0.22.0
idna==3.6
importlib_metadata==7.1.0
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.3
jmespath==1.0.1
lgpio==0.2.2.0
lxml==5.2.1
MarkupSafe==2.1.5
more-itertools==10.2.0
nest_asyncio==1.6.0
numpy==1.26.4
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.1
pigpio==1.78
platformdirs==4.2.0
plotly==5.20.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==1.10.15
PyJWT==2.8.0
pynmea2==1.19.0
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
retrying==1.3.4
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
s3transfer==0.10.1
scipy==1.13.0
seeed-python-reterminal @ git+https://github.com/Seeed-Studio/Seeed_Python_ReTerminal@c7e6a251e9bc3233b79632724e59c2bd4c70507a
simple-salesforce==1.12.6
six==1.16.0
smbus-cffi==0.5.1
smbus2==0.4.3
tenacity==8.2.3
twilio==7.17.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.2
zeep==4.2.1
zipp==3.18.1
